Refactoring a single function to test
